requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boltons==23.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored==2.2.3
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
Django==4.2.7
docx2txt==0.8
eliot==1.15.0
eliot-tree==21.0.0
frozenlist==1.4.0
fsspec==2023.12.0
greenlet==3.0.1
griffe==0.36.9
guardrails-ai==0.2.9
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iso8601==2.1.0
jmespath==1.0.1
joblib==1.3.2
llama-index==0.9.11.post1
lxml==4.9.3
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
openai==1.3.7
orjson==3.9.10
packaging==23.2
pandas==2.1.3
pdf2image==1.16.3
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
pypdf==3.17.1
pyrsistent==0.20.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
reportlab==4.0.7
requests==2.31.0
rich==13.7.0
rstr==3.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
win-unicode-console==0.5
wrapt==1.16.0
yarl==1.9.3
zope.interface==6.1
